Reorganize modules structure
