+++ ctgan/__init__.py
@@ -6,7 +6,7 @@ __author__ = 'MIT Data To AI Lab'
 __email__ = 'dailabmit@gmail.com'
 __version__ = '0.1.1.dev0'
 
-from ctgan.model import CTGANSynthesizer
+from ctgan.synthesizer import CTGANSynthesizer
 
 __all__ = (
     'CTGANSynthesizer',
