--- ctgan/model.py
@@ -1,572 +1,0 @@
-import numpy as np
-import pandas as pd
-import torch
-import torch.utils.data
-from sklearn.mixture import BayesianGaussianMixture
-from torch import optim
-from torch.nn import BatchNorm1d, Dropout, LeakyReLU, Linear, Module, ReLU, Sequential, functional
-
-CONTINUOUS = "continuous"
-CATEGORICAL = "categorical"
-ORDINAL = "ordinal"
-
-
-class DataTransformer(object):
-    """Model continuous columns with a BayesianGMM and normalized to a scalar
-    [0, 1] and a vector.
-    Discrete and ordinal columns are converted to a one-hot vector.
-    """
-
-    def __init__(self, n_clusters=10, eps=0.005):
-        """n_cluster is the upper bound of modes."""
-        self.meta = None
-        self.n_clusters = n_clusters
-        self.eps = eps
-
-    @staticmethod
-    def get_metadata(data, categorical_columns=tuple(), ordinal_columns=tuple()):
-        meta = []
-
-        df = pd.DataFrame(data)
-        for index in df:
-            column = df[index]
-
-            if index in categorical_columns:
-                mapper = column.value_counts().index.tolist()
-                meta.append({
-                    "name": index,
-                    "type": CATEGORICAL,
-                    "size": len(mapper),
-                    "i2s": mapper
-                })
-            elif index in ordinal_columns:
-                value_count = list(dict(column.value_counts()).items())
-                value_count = sorted(value_count, key=lambda x: -x[1])
-                mapper = list(map(lambda x: x[0], value_count))
-                meta.append({
-                    "name": index,
-                    "type": ORDINAL,
-                    "size": len(mapper),
-                    "i2s": mapper
-                })
-            else:
-                meta.append({
-                    "name": index,
-                    "type": CONTINUOUS,
-                    "min": column.min(),
-                    "max": column.max(),
-                })
-
-        return meta
-
-    def fit(self, data, categorical_columns=tuple(), ordinal_columns=tuple()):
-        self.meta = self.get_metadata(data, categorical_columns, ordinal_columns)
-        model = []
-
-        self.output_info = []
-        self.output_dim = 0
-        self.components = []
-        for id_, info in enumerate(self.meta):
-            if info['type'] == CONTINUOUS:
-                gm = BayesianGaussianMixture(
-                    self.n_clusters,
-                    weight_concentration_prior_type='dirichlet_process',
-                    weight_concentration_prior=0.001,
-                    n_init=1
-                )
-                gm.fit(data[:, id_].reshape([-1, 1]))
-                model.append(gm)
-                comp = gm.weights_ > self.eps
-                self.components.append(comp)
-
-                self.output_info += [(1, 'tanh'), (np.sum(comp), 'softmax')]
-                self.output_dim += 1 + np.sum(comp)
-            else:
-                model.append(None)
-                self.components.append(None)
-                self.output_info += [(info['size'], 'softmax')]
-                self.output_dim += info['size']
-
-        self.model = model
-
-    def transform(self, data):
-        values = []
-        for id_, info in enumerate(self.meta):
-            current = data[:, id_]
-            if info['type'] == CONTINUOUS:
-                current = current.reshape([-1, 1])
-
-                means = self.model[id_].means_.reshape((1, self.n_clusters))
-                stds = np.sqrt(self.model[id_].covariances_).reshape((1, self.n_clusters))
-                features = (current - means) / (4 * stds)
-
-                probs = self.model[id_].predict_proba(current.reshape([-1, 1]))
-
-                n_opts = sum(self.components[id_])
-                features = features[:, self.components[id_]]
-                probs = probs[:, self.components[id_]]
-
-                opt_sel = np.zeros(len(data), dtype='int')
-                for i in range(len(data)):
-                    pp = probs[i] + 1e-6
-                    pp = pp / sum(pp)
-                    opt_sel[i] = np.random.choice(np.arange(n_opts), p=pp)
-
-                idx = np.arange((len(features)))
-                features = features[idx, opt_sel].reshape([-1, 1])
-                features = np.clip(features, -.99, .99)
-
-                probs_onehot = np.zeros_like(probs)
-                probs_onehot[np.arange(len(probs)), opt_sel] = 1
-                values += [features, probs_onehot]
-            else:
-                col_t = np.zeros([len(data), info['size']])
-                col_t[np.arange(len(data)), current.astype('int32')] = 1
-                values.append(col_t)
-
-        return np.concatenate(values, axis=1)
-
-    def inverse_transform(self, data, sigmas):
-        data_t = np.zeros([len(data), len(self.meta)])
-
-        st = 0
-        for id_, info in enumerate(self.meta):
-            if info['type'] == CONTINUOUS:
-                u = data[:, st]
-                v = data[:, st + 1:st + 1 + np.sum(self.components[id_])]
-
-                if sigmas is not None:
-                    sig = sigmas[st]
-                    u = np.random.normal(u, sig)
-
-                u = np.clip(u, -1, 1)
-                v_t = np.ones((data.shape[0], self.n_clusters)) * -100
-                v_t[:, self.components[id_]] = v
-                v = v_t
-                st += 1 + np.sum(self.components[id_])
-                means = self.model[id_].means_.reshape([-1])
-                stds = np.sqrt(self.model[id_].covariances_).reshape([-1])
-                p_argmax = np.argmax(v, axis=1)
-                std_t = stds[p_argmax]
-                mean_t = means[p_argmax]
-                tmp = u * 4 * std_t + mean_t
-                data_t[:, id_] = tmp
-
-            else:
-                current = data[:, st:st + info['size']]
-                st += info['size']
-                data_t[:, id_] = np.argmax(current, axis=1)
-
-        return data_t
-
-
-class Discriminator(Module):
-
-    def __init__(self, input_dim, dis_dims, pack=10):
-        super(Discriminator, self).__init__()
-        dim = input_dim * pack
-        self.pack = pack
-        self.packdim = dim
-        seq = []
-        for item in list(dis_dims):
-            seq += [Linear(dim, item), LeakyReLU(0.2), Dropout(0.5)]
-            dim = item
-
-        seq += [Linear(dim, 1)]
-        self.seq = Sequential(*seq)
-
-    def forward(self, input):
-        assert input.size()[0] % self.pack == 0
-        return self.seq(input.view(-1, self.packdim))
-
-
-class Residual(Module):
-    def __init__(self, i, o):
-        super(Residual, self).__init__()
-        self.fc = Linear(i, o)
-        self.bn = BatchNorm1d(o)
-        self.relu = ReLU()
-
-    def forward(self, input):
-        out = self.fc(input)
-        out = self.bn(out)
-        out = self.relu(out)
-        return torch.cat([out, input], dim=1)
-
-
-class Generator(Module):
-    def __init__(self, embedding_dim, gen_dims, data_dim):
-        super(Generator, self).__init__()
-        dim = embedding_dim
-        seq = []
-        for item in list(gen_dims):
-            seq += [Residual(dim, item)]
-            dim += item
-        seq.append(Linear(dim, data_dim))
-        self.seq = Sequential(*seq)
-
-    def forward(self, input):
-        data = self.seq(input)
-        return data
-
-
-def apply_activate(data, output_info):
-    data_t = []
-    st = 0
-    for item in output_info:
-        if item[1] == 'tanh':
-            ed = st + item[0]
-            data_t.append(torch.tanh(data[:, st:ed]))
-            st = ed
-        elif item[1] == 'softmax':
-            ed = st + item[0]
-            data_t.append(functional.gumbel_softmax(data[:, st:ed], tau=0.2))
-            st = ed
-        else:
-            assert 0
-
-    return torch.cat(data_t, dim=1)
-
-
-def random_choice_prob_index(a, axis=1):
-    r = np.expand_dims(np.random.rand(a.shape[1 - axis]), axis=axis)
-    return (a.cumsum(axis=axis) > r).argmax(axis=axis)
-
-
-class Cond(object):
-    def __init__(self, data, output_info):
-        self.model = []
-
-        st = 0
-        skip = False
-        max_interval = 0
-        counter = 0
-        for item in output_info:
-            if item[1] == 'tanh':
-                st += item[0]
-                skip = True
-                continue
-            elif item[1] == 'softmax':
-                if skip:
-                    skip = False
-                    st += item[0]
-                    continue
-
-                ed = st + item[0]
-                max_interval = max(max_interval, ed - st)
-                counter += 1
-                self.model.append(np.argmax(data[:, st:ed], axis=-1))
-                st = ed
-            else:
-                assert 0
-
-        assert st == data.shape[1]
-
-        self.interval = []
-        self.n_col = 0
-        self.n_opt = 0
-        skip = False
-        st = 0
-        self.p = np.zeros((counter, max_interval))
-        for item in output_info:
-            if item[1] == 'tanh':
-                skip = True
-                st += item[0]
-                continue
-            elif item[1] == 'softmax':
-                if skip:
-                    st += item[0]
-                    skip = False
-                    continue
-                ed = st + item[0]
-                tmp = np.sum(data[:, st:ed], axis=0)
-                tmp = np.log(tmp + 1)
-                tmp = tmp / np.sum(tmp)
-                self.p[self.n_col, :item[0]] = tmp
-                self.interval.append((self.n_opt, item[0]))
-                self.n_opt += item[0]
-                self.n_col += 1
-                st = ed
-            else:
-                assert 0
-
-        self.interval = np.asarray(self.interval)
-
-    def sample(self, batch):
-        if self.n_col == 0:
-            return None
-        batch = batch
-        idx = np.random.choice(np.arange(self.n_col), batch)
-
-        vec1 = np.zeros((batch, self.n_opt), dtype='float32')
-        mask1 = np.zeros((batch, self.n_col), dtype='float32')
-        mask1[np.arange(batch), idx] = 1
-        opt1prime = random_choice_prob_index(self.p[idx])
-        opt1 = self.interval[idx, 0] + opt1prime
-        vec1[np.arange(batch), opt1] = 1
-
-        return vec1, mask1, idx, opt1prime
-
-    def sample_zero(self, batch):
-        if self.n_col == 0:
-            return None
-        vec = np.zeros((batch, self.n_opt), dtype='float32')
-        idx = np.random.choice(np.arange(self.n_col), batch)
-        for i in range(batch):
-            col = idx[i]
-            pick = int(np.random.choice(self.model[col]))
-            vec[i, pick + self.interval[col, 0]] = 1
-        return vec
-
-
-def cond_loss(data, output_info, c, m):
-    loss = []
-    st = 0
-    st_c = 0
-    skip = False
-    for item in output_info:
-        if item[1] == 'tanh':
-            st += item[0]
-            skip = True
-
-        elif item[1] == 'softmax':
-            if skip:
-                skip = False
-                st += item[0]
-                continue
-
-            ed = st + item[0]
-            ed_c = st_c + item[0]
-            tmp = functional.cross_entropy(
-                data[:, st:ed],
-                torch.argmax(c[:, st_c:ed_c], dim=1),
-                reduction='none'
-            )
-            loss.append(tmp)
-            st = ed
-            st_c = ed_c
-
-        else:
-            assert 0
-
-    loss = torch.stack(loss, dim=1)
-
-    return (loss * m).sum() / data.size()[0]
-
-
-class Sampler(object):
-    """docstring for Sampler."""
-
-    def __init__(self, data, output_info):
-        super(Sampler, self).__init__()
-        self.data = data
-        self.model = []
-        self.n = len(data)
-
-        st = 0
-        skip = False
-        for item in output_info:
-            if item[1] == 'tanh':
-                st += item[0]
-                skip = True
-            elif item[1] == 'softmax':
-                if skip:
-                    skip = False
-                    st += item[0]
-                    continue
-
-                ed = st + item[0]
-                tmp = []
-                for j in range(item[0]):
-                    tmp.append(np.nonzero(data[:, st + j])[0])
-
-                self.model.append(tmp)
-                st = ed
-            else:
-                assert 0
-
-        assert st == data.shape[1]
-
-    def sample(self, n, col, opt):
-        if col is None:
-            idx = np.random.choice(np.arange(self.n), n)
-            return self.data[idx]
-
-        idx = []
-        for c, o in zip(col, opt):
-            idx.append(np.random.choice(self.model[c][o]))
-
-        return self.data[idx]
-
-
-def calc_gradient_penalty(netD, real_data, fake_data, device='cpu', pac=10, lambda_=10):
-
-    alpha = torch.rand(real_data.size(0) // pac, 1, 1, device=device)
-    alpha = alpha.repeat(1, pac, real_data.size(1))
-    alpha = alpha.view(-1, real_data.size(1))
-
-    interpolates = alpha * real_data + ((1 - alpha) * fake_data)
-
-    disc_interpolates = netD(interpolates)
-
-    gradients = torch.autograd.grad(
-        outputs=disc_interpolates, inputs=interpolates,
-        grad_outputs=torch.ones(disc_interpolates.size(), device=device),
-        create_graph=True, retain_graph=True, only_inputs=True
-    )[0]
-
-    gradient_penalty = ((
-        gradients.view(-1, pac * real_data.size(1)).norm(2, dim=1) - 1
-    ) ** 2).mean() * lambda_
-
-    return gradient_penalty
-
-
-class CTGANSynthesizer(object):
-    """docstring for IdentitySynthesizer."""
-
-    def __init__(self, embedding_dim=128, gen_dim=(256, 256), dis_dim=(256, 256),
-                 l2scale=1e-6, batch_size=500, epochs=300):
-
-        self.embedding_dim = embedding_dim
-        self.gen_dim = gen_dim
-        self.dis_dim = dis_dim
-
-        self.l2scale = l2scale
-        self.batch_size = batch_size
-        self.epochs = epochs
-        self.device = torch.device("cuda:0" if torch.cuda.is_available() else "cpu")
-
-    def fit(self, train_data, categorical_columns=tuple(), ordinal_columns=tuple()):
-
-        self.transformer = DataTransformer()
-        self.transformer.fit(train_data, categorical_columns, ordinal_columns)
-        train_data = self.transformer.transform(train_data)
-
-        data_sampler = Sampler(train_data, self.transformer.output_info)
-
-        data_dim = self.transformer.output_dim
-        self.cond_generator = Cond(train_data, self.transformer.output_info)
-
-        self.generator = Generator(
-            self.embedding_dim + self.cond_generator.n_opt,
-            self.gen_dim,
-            data_dim
-        ).to(self.device)
-
-        discriminator = Discriminator(
-            data_dim + self.cond_generator.n_opt,
-            self.dis_dim
-        ).to(self.device)
-
-        optimizerG = optim.Adam(
-            self.generator.parameters(), lr=2e-4, betas=(0.5, 0.9),
-            weight_decay=self.l2scale
-        )
-        optimizerD = optim.Adam(discriminator.parameters(), lr=2e-4, betas=(0.5, 0.9))
-
-        assert self.batch_size % 2 == 0
-        mean = torch.zeros(self.batch_size, self.embedding_dim, device=self.device)
-        std = mean + 1
-
-        steps_per_epoch = len(train_data) // self.batch_size
-        for i in range(self.epochs):
-            for id_ in range(steps_per_epoch):
-                fakez = torch.normal(mean=mean, std=std)
-
-                condvec = self.cond_generator.sample(self.batch_size)
-                if condvec is None:
-                    c1, m1, col, opt = None, None, None, None
-                    real = data_sampler.sample(self.batch_size, col, opt)
-                else:
-                    c1, m1, col, opt = condvec
-                    c1 = torch.from_numpy(c1).to(self.device)
-                    m1 = torch.from_numpy(m1).to(self.device)
-                    fakez = torch.cat([fakez, c1], dim=1)
-
-                    perm = np.arange(self.batch_size)
-                    np.random.shuffle(perm)
-                    real = data_sampler.sample(self.batch_size, col[perm], opt[perm])
-                    c2 = c1[perm]
-
-                fake = self.generator(fakez)
-                fakeact = apply_activate(fake, self.transformer.output_info)
-
-                real = torch.from_numpy(real.astype('float32')).to(self.device)
-
-                if c1 is not None:
-                    fake_cat = torch.cat([fakeact, c1], dim=1)
-                    real_cat = torch.cat([real, c2], dim=1)
-                else:
-                    real_cat = real
-                    fake_cat = fake
-
-                y_fake = discriminator(fake_cat)
-                y_real = discriminator(real_cat)
-
-                loss_d = -(torch.mean(y_real) - torch.mean(y_fake))
-                pen = calc_gradient_penalty(discriminator, real_cat, fake_cat, self.device)
-
-                optimizerD.zero_grad()
-                pen.backward(retain_graph=True)
-                loss_d.backward()
-                optimizerD.step()
-
-                fakez = torch.normal(mean=mean, std=std)
-                condvec = self.cond_generator.sample(self.batch_size)
-
-                if condvec is None:
-                    c1, m1, col, opt = None, None, None, None
-                else:
-                    c1, m1, col, opt = condvec
-                    c1 = torch.from_numpy(c1).to(self.device)
-                    m1 = torch.from_numpy(m1).to(self.device)
-                    fakez = torch.cat([fakez, c1], dim=1)
-
-                fake = self.generator(fakez)
-                fakeact = apply_activate(fake, self.transformer.output_info)
-
-                if c1 is not None:
-                    y_fake = discriminator(torch.cat([fakeact, c1], dim=1))
-                else:
-                    y_fake = discriminator(fakeact)
-
-                if condvec is None:
-                    cross_entropy = 0
-                else:
-                    cross_entropy = cond_loss(fake, self.transformer.output_info, c1, m1)
-
-                loss_g = -torch.mean(y_fake) + cross_entropy
-
-                optimizerG.zero_grad()
-                loss_g.backward()
-                optimizerG.step()
-
-            print("Epoch %d, Loss G: %.4f, Loss D: %.4f" %
-                  (i + 1, loss_g.detach().cpu(), loss_d.detach().cpu()))
-
-    def sample(self, n):
-
-        output_info = self.transformer.output_info
-        steps = n // self.batch_size + 1
-        data = []
-        for i in range(steps):
-            mean = torch.zeros(self.batch_size, self.embedding_dim)
-            std = mean + 1
-            fakez = torch.normal(mean=mean, std=std).to(self.device)
-
-            condvec = self.cond_generator.sample_zero(self.batch_size)
-            if condvec is None:
-                pass
-            else:
-                c1 = condvec
-                c1 = torch.from_numpy(c1).to(self.device)
-                fakez = torch.cat([fakez, c1], dim=1)
-
-            fake = self.generator(fakez)
-            fakeact = apply_activate(fake, output_info)
-            data.append(fakeact.detach().cpu().numpy())
-
-        data = np.concatenate(data, axis=0)
-        data = data[:n]
-        return self.transformer.inverse_transform(data, None)
