+++ ctgan/transformer.py
@@ -1,0 +1,174 @@
+import numpy as np
+import pandas as pd
+from sklearn.mixture import BayesianGaussianMixture
+
+CONTINUOUS = "continuous"
+CATEGORICAL = "categorical"
+ORDINAL = "ordinal"
+
+
+class DataTransformer(object):
+    """Data Transformer.
+
+    Model continuous columns with a BayesianGMM and normalized to a scalar
+    [0, 1] and a vector.
+    Discrete and ordinal columns are converted to a one-hot vector.
+
+    Args:
+        n_cluster (int):
+            Number of modes.
+        eps (float):
+            Epsilon value.
+    """
+
+    def __init__(self, n_clusters=10, eps=0.005):
+        self.n_clusters = n_clusters
+        self.eps = eps
+
+    @staticmethod
+    def get_metadata(data, categorical_columns, ordinal_columns):
+        """Generate the dataset metadata.
+
+        Args:
+            categorical_columns (list[int]):
+                Indexes of the categorical columns.
+            ordinal_columns (list[int]):
+                Indexes of the categorical columns.
+
+        Returns:
+            list:
+                Generated metadata.
+        """
+        meta = []
+
+        df = pd.DataFrame(data)
+        for index in df:
+            column = df[index]
+
+            if index in categorical_columns:
+                mapper = column.value_counts().index.tolist()
+                meta.append({
+                    "name": index,
+                    "type": CATEGORICAL,
+                    "size": len(mapper),
+                    "i2s": mapper
+                })
+            elif index in ordinal_columns:
+                value_count = list(dict(column.value_counts()).items())
+                value_count = sorted(value_count, key=lambda x: -x[1])
+                mapper = list(map(lambda x: x[0], value_count))
+                meta.append({
+                    "name": index,
+                    "type": ORDINAL,
+                    "size": len(mapper),
+                    "i2s": mapper
+                })
+            else:
+                meta.append({
+                    "name": index,
+                    "type": CONTINUOUS,
+                    "min": column.min(),
+                    "max": column.max(),
+                })
+
+        return meta
+
+    def fit(self, data, categorical_columns=tuple(), ordinal_columns=tuple()):
+        self.meta = self.get_metadata(data, categorical_columns, ordinal_columns)
+        models = []
+
+        self.output_info = []
+        self.output_dim = 0
+        self.components = []
+        for id_, info in enumerate(self.meta):
+            if info['type'] == CONTINUOUS:
+                gm = BayesianGaussianMixture(
+                    self.n_clusters,
+                    weight_concentration_prior_type='dirichlet_process',
+                    weight_concentration_prior=0.001,
+                    n_init=1
+                )
+                gm.fit(data[:, id_].reshape([-1, 1]))
+                models.append(gm)
+                components = gm.weights_ > self.eps
+                self.components.append(components)
+
+                self.output_info += [(1, 'tanh'), (np.sum(components), 'softmax')]
+                self.output_dim += 1 + np.sum(components)
+            else:
+                models.append(None)
+                self.components.append(None)
+                self.output_info += [(info['size'], 'softmax')]
+                self.output_dim += info['size']
+
+        self.models = models
+
+    def transform(self, data):
+        values = []
+        for id_, info in enumerate(self.meta):
+            current = data[:, id_]
+            if info['type'] == CONTINUOUS:
+                current = current.reshape([-1, 1])
+
+                means = self.models[id_].means_.reshape((1, self.n_clusters))
+                stds = np.sqrt(self.models[id_].covariances_).reshape((1, self.n_clusters))
+                features = (current - means) / (4 * stds)
+
+                probs = self.models[id_].predict_proba(current.reshape([-1, 1]))
+
+                n_opts = sum(self.components[id_])
+                features = features[:, self.components[id_]]
+                probs = probs[:, self.components[id_]]
+
+                opt_sel = np.zeros(len(data), dtype='int')
+                for i in range(len(data)):
+                    pp = probs[i] + 1e-6
+                    pp = pp / sum(pp)
+                    opt_sel[i] = np.random.choice(np.arange(n_opts), p=pp)
+
+                idx = np.arange((len(features)))
+                features = features[idx, opt_sel].reshape([-1, 1])
+                features = np.clip(features, -.99, .99)
+
+                probs_onehot = np.zeros_like(probs)
+                probs_onehot[np.arange(len(probs)), opt_sel] = 1
+                values += [features, probs_onehot]
+            else:
+                col_t = np.zeros([len(data), info['size']])
+                col_t[np.arange(len(data)), current.astype('int32')] = 1
+                values.append(col_t)
+
+        return np.concatenate(values, axis=1)
+
+    def inverse_transform(self, data, sigmas):
+        data_t = np.zeros([len(data), len(self.meta)])
+
+        st = 0
+        for id_, info in enumerate(self.meta):
+            if info['type'] == CONTINUOUS:
+                u = data[:, st]
+                v = data[:, st + 1:st + 1 + np.sum(self.components[id_])]
+
+                if sigmas is not None:
+                    sig = sigmas[st]
+                    u = np.random.normal(u, sig)
+
+                u = np.clip(u, -1, 1)
+                v_t = np.ones((data.shape[0], self.n_clusters)) * -100
+                v_t[:, self.components[id_]] = v
+                v = v_t
+                st += 1 + np.sum(self.components[id_])
+                means = self.models[id_].means_.reshape([-1])
+                stds = np.sqrt(self.models[id_].covariances_).reshape([-1])
+                p_argmax = np.argmax(v, axis=1)
+                std_t = stds[p_argmax]
+                mean_t = means[p_argmax]
+                tmp = u * 4 * std_t + mean_t
+                data_t[:, id_] = tmp
+
+            else:
+                current = data[:, st:st + info['size']]
+                st += info['size']
+                data_t[:, id_] = np.argmax(current, axis=1)
+
+        return data_t
